+++ TP1_Integrador/src/DAO/MySQLClienteDAO.java
@@ -12,7 +12,7 @@ public class MySQLClienteDAO implements UsuarioDAO{
 		this.conn=conn;
 	}
 	@Override
-	public void insertar(Long id, String nombre, String email ) throws SQLException {
+	public void insertar(INT id, String nombre, String email ) throws SQLException {
 		String table= "INSERT INTO Cliente(idCliente,email,nombre) VALUES (?, ?, ?)";
 		PreparedStatement ps = conn.prepareStatement(insert);
 		ps.setInt(1, id);
@@ -25,17 +25,17 @@ public class MySQLClienteDAO implements UsuarioDAO{
 	}
 
 	@Override
-	public void delete(int id) {
+	public void delete(INT id) {
 		
 	}
 
 	@Override
-	public void update(Long id, String nombre, String email ) throws SQLException {
+	public void update(INT id, String nombre, String email ) throws SQLException {
 	
 	}
 
 	@Override
-	public Usuario getById(int id){
+	public Usuario getById(INT id){
 
 	}
