+++ TP1_Integrador/src/DAO/MySQLFacturaDAO.java
@@ -2,7 +2,7 @@ package DAO;
 
 public class MySQLFacturaDAO implements FacturaDao  {
 
-	public void insertar(Long id, String nombre, String email ) {
+	public void insertar(INT id, String nombre, String email ) {
 		System.out.println(id +"-"+nombre+"-"+email);
 		
 	}
