+++ TP1_Integrador/src/DAO/ProductoDao.java
@@ -4,13 +4,13 @@ import java.sql.SQLException;
 import java.util.List;
 
 public interface ProductoDAO {
-    public void insert(Long id, String nombre, float valor ) throws SQLException;
+    public void insert(INT id, String nombre, float valor ) throws SQLException;
 
-    public void update(Long id, String nombre, float valor ) throws SQLException;
+    public void update(INT id, String nombre, float valor ) throws SQLException;
 
-    public void delete(int id);
+    public void delete(INT id);
 
-    public Producto getById(int id);
+    public Producto getById(INT id);
 
     public List<Producto> getAll();
 }
