+++ TP1_Integrador/src/DAO/UsuarioDAO.java
@@ -5,13 +5,13 @@ import java.sql.SQLException;
 public interface UsuarioDAO {
 
 
-    public void insert(Long id, String nombre, String email ) throws SQLException;
+    public void insert(INT id, String nombre, String email ) throws SQLException;
 
-    public void update(Long id, String nombre, String email ) throws SQLException;
+    public void update(INT id, String nombre, String email ) throws SQLException;
 
-    public void delete(Long id);
+    public void delete(INT id);
 
-    public Usuario getById(Long id);
+    public Usuario getById(INT id);
 
     public List<Usuario> getAll();
